feat: Add fish completion for flyctl
This commit adds fish shell completion for the flyctl command. The completion logic is implemented in the `flyctl.fish` file. The completion script works by extracting command line arguments and performing completion requests using the `__complete` function. The completion results are then processed and formatted to be displayed to the user.

Additionally, this commit adds an alias for `flyctl` to `fly` in the `config.fish` file. This allows users to use the shortened `fly` command instead of typing `flyctl`.