Update fish_variables
This commit updates the fish_variables file in the fish directory. Specifically, it modifies the export values of ABBR_TIPS_PROMPT, ABBR_TIPS_REGEXES, and __ABBR_TIPS_KEYS in the fish_variables file. It also changes the export value of __ABBR_TIPS_VALUES. These modifications are made to improve the functionality of the variables in the fish shell configuration.