feat: Add 'serve' subcommand completion
This commit adds completion for the 'serve' subcommand in the deno fish completions file. The 'serve' subcommand allows running a server. The completion includes options such as '--no-check' to skip type-checking modules, '--import-map' to load an import map file, '--node-modules-dir' to enable or disable the use of a local node_modules folder, and '--vendor' to enable or disable the use of a local vendor folder for remote modules and node_modules folders. It also includes permission completions such as '--allow-read', '--allow-write', '--allow-net', and more. For a full list of options and permissions, please see the documentation at https://deno.land/manual@v1.43.2/.